RSMS-1104: Cast isotope report 'auth_limit' field to Float; not int.
This field needs to support non-integer values.

+++ rsms/src/includes/modules/radiation/dto/RadReportDTO.php
@@ -104,7 +104,7 @@ class RadReportDTO
 	}
 
 	public function getAuth_limit(){
-		return (int) $this->auth_limit;
+		return (float) $this->auth_limit;
 	}
 
 	public function setAuth_limit($auth_limit){
